Add files via upload
Lidar module for nandi project. This uses a Slamtec C1 Lidar. The hardware comes very poorly documented. No documented python serial / lidar libraries will work with this hardare. The only way to use this is to get the Slamtec sdk code, and compile it locally. Then use this as an executable to generate distance data. Once that is done, we can use python to ingest the data and further use it. This script begins to do that. The output file contains the Lidar readout.