Add files via upload
ESP32 .ino file that scans for BLE devices, connects to WiFi network, and to a MQTT Broker on the same network, and pass messages to the broker about BL devices identified. Code cannot distinguish between BLE beacons and BL devices in proximity. Code will disregard all devices that do not have a device name.

Once burned in, the code will do the following:
1. connect to the wifi network (arduino_secrets.h in the same directory will have creds)
2. connect to the MQTT broker (again, access creds in arduino_secrets.h)
3. initiate BLE scan in ESP32
4. run a scan every 2 seconds, and send a message to MQTT broker.
5. Each message is a serialized json string, with the following components:
    5a. sender: (name of the ESP32 device - set it in the code)
    5b. beacon: (name of the BL/BLE device)
    5c. TXPower - almost always set to zero (0) - but may come in handy in complex situations
    5d. RSSI : Received Signal Strength Indicator in dB. Lower the number, worse the signal (the numbers are below zero)
    5e. MAC Address of the device (tip - if your BLE doesn't have a device name, then use MAC address as a name)