Add files via upload
Sketch is to be used with ESP32 dev board (works on at least two versions). It defines BLE server, service and characteristics. It specifically allows the addition of more than 4 characteristics and a descriptor or two for each. Code will be adjusted to provide sufficient descriptors (BLE2901 and BLE2902) for each characteristic. Currently, the code allows for notification of four of the seven defined characteristics, and an alternate label for 4 of the 7 characteristics. Tested with LightBlue client.